backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.0
requests==2.31.0
python-multipart==0.0.6
pydantic==2.4.2
pydantic-settings==2.0.3
slowapi==0.1.8
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
sqlalchemy==2.0.23
alembic==1.12.1
python-jose[cryptography]==3.3.0 